Fixes & Optimizations
-Fix wrong shaders/material info for world mesh when changing level

-Use more than one temporary vertex buffer to avoid sending more than necessary data to gpu and stabilize frame times + avoid vram fragmentation

-Optimize a bit world mesh obtaining per frame

+++ D3D11Engine/D2DSettingsDialog.cpp
@@ -189,7 +189,7 @@ XRESULT D2DSettingsDialog::InitControls() {
 	fpsLimitSlider->SetDisplayValues( fpsValues );
 
 	// Fix the fps value
-	fpsLimitSlider->SetValue( (float)Engine::GAPI->GetRendererState().RendererSettings.FpsLimit );
+	fpsLimitSlider->SetValue( Engine::GAPI->GetRendererState().RendererSettings.FpsLimit );
 
 	// Next column
 	SV_Label* outdoorVobsDDLabel = new SV_Label( MainView, MainPanel );
@@ -355,7 +355,7 @@ XRESULT D2DSettingsDialog::InitControls() {
 }
 
 void D2DSettingsDialog::FpsLimitSliderChanged( SV_Slider* sender, void* userdata ) {
-	int newValue = (int)sender->GetValue();
+	int newValue = sender->GetValue();
 	Engine::GAPI->GetRendererState().RendererSettings.FpsLimit = newValue <= 25 ? 0 : newValue;
 }
